pyproject.toml
[tool.poetry]
name = "caesar-cipher"
version = "0.1.0"
description = ""
authors = ["Leo Kukharau <kuhorev.leonid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.0.0"
english-words = "^1.0.3"
pyenchant = "^3.1.1"


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
